Mopho 2: In RandomizationOptions, renamed allowChoiceForParam() to setChoiceIsAllowedForParam() and added a bool shouldBeAllowed argument to its signature. Eliminated choiceIsForbiddenForParam() and forbidChoiceForParam(). Declared and defined clearAllowedChoicesForParam() and allowAllChoicesForParam(). Updated AllowChoiceToggles to the revised functions and refined the logic in buttonClicked().

+++ MophoGUI/MophoGUI 2.0/Source/randomization/rndm_0_comp_AllowChoiceToggles.h
@@ -14,7 +14,7 @@ class AllowChoiceToggles :
 	uint8 paramIndex;
 	UnexposedParameters* unexposedParams;
 	std::vector<std::unique_ptr<ToggleButton>> allowedChoiceToggles;
-	int numberOfChoices;
+	uint8 numberOfChoices;
 	int numberOfColumns;
 	int numberOfRows;
 	int toggle_w;
